refact: transfer table center to another file
