+++ src/components/domain/table/types.ts
@@ -1,0 +1,9 @@
+export type Dimensions = "XS" | "default";
+
+export type TableResponsiveConfig = {
+  limits: number[];
+  newMatchButton: string;
+  revealCardsButton: string;
+};
+
+export type TableResponsiveConfigs = Record<Dimensions, TableResponsiveConfig>;
